Move code to async_timeout.py
